Restructure server storage code to make more module-private
