+++ src/base/mod.rs
@@ -1,2 +1,6 @@
+mod local_context;
 pub mod message_utils;
+pub mod raft_utils;
 pub mod utils;
+
+pub use local_context::LocalContext;
