+++ src/client/cluster_client.rs
@@ -1,9 +1,9 @@
 use crate::base::message_utils::accessed_inode;
+use crate::base::raft_utils::node_contains_raft_group;
 use crate::base::utils::{finalize_request_without_prefix, LengthPrefixedVec};
+use crate::base::LocalContext;
 use crate::client::PeerClient;
 use crate::generated::*;
-use crate::server::LocalContext;
-use crate::storage::raft_node::node_contains_raft_group;
 use flatbuffers::FlatBufferBuilder;
 use futures_util::future::FutureExt;
 use std::collections::HashMap;
