--- src/server/mod.rs
@@ -1,7 +1,0 @@
-mod fsck_handler;
-mod router;
-mod storage_node;
-mod transaction_coordinator;
-
-pub use storage_node::LocalContext;
-pub use storage_node::Node;
