+++ src/storage/mod.rs
@@ -1,6 +1,9 @@
 mod local;
 mod lock_table;
-pub mod raft_group_manager;
-pub mod raft_node;
+mod message_handlers;
+mod raft_group_manager;
+mod raft_node;
+mod storage_node;
 
 pub use local::ROOT_INODE;
+pub use storage_node::Node;
